requirements.txt
flask==3.1.3
flask-cors==6.0.2
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
gunicorn==21.2.0
